requirements.txt
streamlit==1.24.1
pandas==1.5.3
PyYAML==6.0
bcrypt==4.0.1
openpyxl==3.0.10
XlsxWriter==3.0.3
python-pptx>=0.6.21